// frontend/src/api/generated/models/strategyResponse.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * vibe-quant API
 * OpenAPI spec version: 0.1.0
 */
import type { StrategyResponseDslConfig } from './strategyResponseDslConfig';

export interface StrategyResponse {
  id: number;
  name: string;
  description: string | null;
  dsl_config: StrategyResponseDslConfig;
  strategy_type: string | null;
  created_at: string;
  updated_at: string;
  is_active: boolean;
  version: number;
}